+++ amethyst_ui/src/bundle.rs
@@ -18,11 +18,11 @@ use super::*;
 /// The generic types A and B represent the A and B generic parameter of the InputHandler<A,B>.
 ///
 /// Will fail with error 'No resource with the given id' if the InputBundle is not added.
-pub struct UiBundle<A, B> {
-    _marker: PhantomData<(A, B)>,
+pub struct UiBundle<A, B, C = NoCustomUi> {
+    _marker: PhantomData<(A, B, C)>,
 }
 
-impl<A, B> UiBundle<A, B> {
+impl<A, B, C> UiBundle<A, B, C> {
     /// Create a new UI bundle
     pub fn new() -> Self {
         UiBundle {
@@ -31,14 +31,20 @@ impl<A, B> UiBundle<A, B> {
     }
 }
 
-impl<'a, 'b, A, B> SystemBundle<'a, 'b> for UiBundle<A, B>
+impl<'a, 'b, A, B, C> SystemBundle<'a, 'b> for UiBundle<A, B, C>
 where
     A: Send + Sync + Eq + Hash + Clone + 'static,
     B: Send + Sync + Eq + Hash + Clone + 'static,
+    C: ToNativeWidget,
 {
     fn build(self, builder: &mut DispatcherBuilder<'a, 'b>) -> Result<()> {
         builder.add(
-            UiLoaderSystem::<AudioFormat, TextureFormat, FontFormat>::default(),
+            UiLoaderSystem::<
+                AudioFormat,
+                TextureFormat,
+                FontFormat,
+                <C as ToNativeWidget>::PrefabData,
+            >::default(),
             "ui_loader",
             &[],
         );